Refactor PRReviewAgent into modular components
